Start on replacing Space Invaders audio with samples
This commit adds sample data to the Space Invaders MRA. The data
was taken from the WAV files at
https://www.classicgaming.cc/classics/space-invaders/sounds

The sounds aren't working - I think the config strings are
wrong, but I don't have time to figure it out right now. For
example, the only sound that comes out is when an invader is
destroyed, but that sounds like one of the marching sounds.