Use samples for sounds
This commit fixes up the sound issues from the previous commit
so that the sounds are coming from the sample files found at
https://www.classicgaming.cc/classics/space-invaders/sounds

Note that I flipped the shot and invader killed sounds, as I
think that they were reversed. Also, there's no extra base sound,
as that's not on the page. Finally, I editted the low-pitch mystery
ship sound so that it would repeat properly.